chore: reorganize uplot scatter plot files
