+++ webui/react/src/components/UPlot/UPlotScatter/UPlotScatter.utils.ts
@@ -3,7 +3,7 @@ import uPlot from 'uplot';
 import { Range } from 'types';
 import { rgba2str, rgbaFromGradient, str2rgba } from 'utils/color';
 
-import { UPlotAxisSplits, UPlotData } from './types';
+import { UPlotAxisSplits, UPlotData } from '../types';
 
 export const X_INDEX = 0;
 export const Y_INDEX = 1;
